LayoutTests/platform/blackberry/accessibility/label-for-control-hittest-expected.txt

 This is a lot of repeated text that when a hit test is performed, will return the corresponding checkbox control, because this text is only a label. This is a lot of repeated text that when a hit test is performed, will return the corresponding checkbox control, because this text is only a label. This is a lot of repeated text that when a hit test is performed, will return the corresponding checkbox control, because this text is only a label. This is a lot of repeated text that when a hit test is performed, will return the corresponding checkbox control, because this text is only a label. This is a lot of repeated text that when a hit test is performed, will return the corresponding checkbox control, because this text is only a label.
This tests that a hit test on a label of a control will return the control (instead of nothing).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL control.role should be AXRole: AXCheckBox. Was .
PASS successfullyParsed is true

TEST COMPLETE
